added API endpoint and HTTP method to API method doc blocks
The comment follows the same style as in the scopes class, making it easy for devlopers to see which endpoints they use and therefore which scopes they need to request access to. Developers can scan the methods they use to see the endpoints and then look for the corresponding endpoint in the scopes class to deterimine which scopes they need access to.